REFECTORING:: send add all, folder change
