+++ src/modal/aldetail.jsx
